Create README.md
Initial Commit. 

This Repo is serving as a "Placeholder" temporarily. I'll be using it to set up projects, save some random files, ranging from html/css/js to PHP to Python to GDscript etc....

Will Add Updates When i actually Begin work on "Mumblr_Memo" Which will be a.... 

-> Password Protected/Encrypted Markdown Note Taking  Web app + code editor + image gallery.  With Intention To Release On Android As Well. 

LATER!